depressed/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
py-eureka-client==0.7.4